fix: Post Module part 2:
- fixed magic strings (added every constants to
libs/shared/constants)
- Added minioservice and separate logic about adding, deleting
objects from storage to this service
- deleted unnecesary modules in innogram/app.module.ts
- changed name for swagger create posts body decorator
- changed typing params from Type["field"] to just string
- separated getPaginationParams logic to utils
- delegated checking if images are exist to images service
- ParseUUIDPipe inbuild instead of custom
- changed shape of FILE URL in db, now only path to resource without
  host and bucket name (before
  http://localhost:3000/inngoram/posts/name.png, now just posts/name.png)
- added .vscode/settings.json with option to work on LF mode
  with this project on VSC
- added .gitattributes eol=lf to make sure that when making commits,
  files are in lf mode

+++ backend/apps/innogram/src/image/image.service.ts
@@ -1,8 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { FileService } from '../file/file.service';
-import { FileSubdirectory } from '../file/enum/file.enum';
+import { type Image, FileSubdirectory, hasItems } from '@app/shared';
 import { ImageRepository } from './image.repository';
-import { type Image } from '@app/shared';
 
 @Injectable()
 export class ImageService {
@@ -16,6 +15,10 @@ export class ImageService {
     subdirectory: FileSubdirectory,
     postId?: string,
   ) {
+    if (!hasItems(files)) {
+      return [];
+    }
+
     const uploadedFiles = await this.fileService.uploadFiles(
       files,
       subdirectory,
@@ -30,11 +33,12 @@ export class ImageService {
   }
 
   deleteImages(images: Image[]) {
+    if (!hasItems(images)) {
+      return Promise.resolve();
+    }
+
     const imageIds = images.map((image) => image.id);
-    const filesToDelete = images.map((image) => ({
-      id: image.file.id,
-      url: image.file.url,
-    }));
+    const filesToDelete = images.map((image) => image.file);
 
     return Promise.all([
       this.fileService.deleteFiles(filesToDelete),
